requirements.txt
beautifulsoup4==4.9.3
tabula-py==2.2.0
